changed all dimensionality to type unsigned int

+++ include/algorithms.h
@@ -5,9 +5,9 @@
 #include <vector>
 
 #include "filtration_algorithms.h"
-    
+
 struct dimension_filtration : public filtration_algorithm_t {
-    virtual inline value_t compute_filtration(unsigned short dimension, const directed_flag_complex_cell_t& cell,
+    virtual inline value_t compute_filtration(unsigned int dimension, const directed_flag_complex_cell_t& cell,
                                     const filtered_directed_graph_t& graph,
                                     const value_t* boundary_filtration) const {
         return (dimension);
@@ -18,7 +18,7 @@ virtual inline bool overwrite_edge_filtration() const { return true; }
 };
 
 struct zero_filtration : public filtration_algorithm_t {
-    virtual inline value_t compute_filtration(unsigned short dimension, const directed_flag_complex_cell_t& cell,
+    virtual inline value_t compute_filtration(unsigned int dimension, const directed_flag_complex_cell_t& cell,
                                     const filtered_directed_graph_t& graph,
                                     const value_t* boundary_filtration) const {
         return (0.0f);
@@ -29,7 +29,7 @@ virtual inline bool overwrite_edge_filtration() const { return true; }
 };
 
 struct max_filtration : public filtration_algorithm_t {
-    virtual inline value_t compute_filtration(unsigned short dimension, const directed_flag_complex_cell_t& cell,
+    virtual inline value_t compute_filtration(unsigned int dimension, const directed_flag_complex_cell_t& cell,
                                     const filtered_directed_graph_t& graph,
                                     const value_t* boundary_filtration) const {
         return (max(boundary_filtration, 0, dimension));
@@ -37,7 +37,7 @@ struct max_filtration : public filtration_algorithm_t {
 };
 
 struct max3_filtration : public filtration_algorithm_t {
-    virtual inline value_t compute_filtration(unsigned short dimension, const directed_flag_complex_cell_t& cell,
+    virtual inline value_t compute_filtration(unsigned int dimension, const directed_flag_complex_cell_t& cell,
                                     const filtered_directed_graph_t& graph,
                                     const value_t* boundary_filtration) const {
         return (max(boundary_filtration, 0, dimension));
@@ -45,7 +45,7 @@ struct max3_filtration : public filtration_algorithm_t {
 };
 
 struct max_plus_one_filtration : public filtration_algorithm_t {
-    virtual inline value_t compute_filtration(unsigned short dimension, const directed_flag_complex_cell_t& cell,
+    virtual inline value_t compute_filtration(unsigned int dimension, const directed_flag_complex_cell_t& cell,
                                     const filtered_directed_graph_t& graph,
                                     const value_t* boundary_filtration) const {
         return (max(boundary_filtration, 0, dimension)+1.0f);
@@ -53,7 +53,7 @@ struct max_plus_one_filtration : public filtration_algorithm_t {
 };
 
 struct product_filtration : public filtration_algorithm_t {
-    virtual inline value_t compute_filtration(unsigned short dimension, const directed_flag_complex_cell_t& cell,
+    virtual inline value_t compute_filtration(unsigned int dimension, const directed_flag_complex_cell_t& cell,
                                     const filtered_directed_graph_t& graph,
                                     const value_t* boundary_filtration) const {
         return (product(boundary_filtration, 0, dimension));
@@ -61,7 +61,7 @@ struct product_filtration : public filtration_algorithm_t {
 };
 
 struct sum_filtration : public filtration_algorithm_t {
-    virtual inline value_t compute_filtration(unsigned short dimension, const directed_flag_complex_cell_t& cell,
+    virtual inline value_t compute_filtration(unsigned int dimension, const directed_flag_complex_cell_t& cell,
                                     const filtered_directed_graph_t& graph,
                                     const value_t* boundary_filtration) const {
         return (sum(boundary_filtration, 0, dimension));
@@ -69,7 +69,7 @@ struct sum_filtration : public filtration_algorithm_t {
 };
 
 struct pmean_filtration : public filtration_algorithm_t {
-    virtual inline value_t compute_filtration(unsigned short dimension, const directed_flag_complex_cell_t& cell,
+    virtual inline value_t compute_filtration(unsigned int dimension, const directed_flag_complex_cell_t& cell,
                                     const filtered_directed_graph_t& graph,
                                     const value_t* boundary_filtration) const {
         return (max(boundary_filtration, 0, dimension)+pow(
@@ -84,7 +84,7 @@ struct pmean_filtration : public filtration_algorithm_t {
 };
 
 struct pmoment_filtration : public filtration_algorithm_t {
-    virtual inline value_t compute_filtration(unsigned short dimension, const directed_flag_complex_cell_t& cell,
+    virtual inline value_t compute_filtration(unsigned int dimension, const directed_flag_complex_cell_t& cell,
                                     const filtered_directed_graph_t& graph,
                                     const value_t* boundary_filtration) const {
         return (max(boundary_filtration, 0, dimension)+pow(
@@ -99,29 +99,29 @@ struct pmoment_filtration : public filtration_algorithm_t {
 };
 
 struct remove_edges_filtration : public filtration_algorithm_t {
-    virtual inline value_t compute_filtration(unsigned short dimension, const directed_flag_complex_cell_t& cell,
+    virtual inline value_t compute_filtration(unsigned int dimension, const directed_flag_complex_cell_t& cell,
                                     const filtered_directed_graph_t& graph,
                                     const value_t* boundary_filtration) const {
-        
+
                     if (dimension == 0) {
                         return (0.0f);
                     }
-                    
+
                     if (dimension == 1) {
-                        
+
             std::cerr << "Please provide a graph with random weights on the edges." << std::endl;
             exit(-1);
-    
+
                     }
                     return (max(boundary_filtration, 0, dimension));
     }
 };
 
 struct vertex_degree_filtration : public filtration_algorithm_t {
-    virtual inline value_t compute_filtration(unsigned short dimension, const directed_flag_complex_cell_t& cell,
+    virtual inline value_t compute_filtration(unsigned int dimension, const directed_flag_complex_cell_t& cell,
                                     const filtered_directed_graph_t& graph,
                                     const value_t* boundary_filtration) const {
-        
+
                     if (dimension == 0) {
                         return (-1.0f*(graph.outdegrees[(int)((cell[(int)((0.0f))]))]+graph.indegrees[(int)((cell[(int)((0.0f))]))]));
                     }
@@ -148,4 +148,3 @@ if (algorithm == "vertex_degree") return new vertex_degree_filtration();
     }
     std::vector<std::string> custom_filtration_computer{ "dimension", "zero", "max", "max3", "max_plus_one", "product", "sum", "pmean", "pmoment", "remove_edges", "vertex_degree" };
     #endif // FLAGSER_ALGORITHMS_H
-
