+++ include/persistence.h
@@ -479,15 +479,15 @@ public:
 
 	void set_print_betti_numbers(bool print_betti_numbers) { print_betti_numbers_to_console = print_betti_numbers; }
 
-	void compute_persistence(unsigned short min_dimension = 0,
-	                         unsigned short max_dimension = std::numeric_limits<unsigned short>::max(), bool check_euler_characteristic = true) {
+	void compute_persistence(unsigned int min_dimension = 0,
+	                         unsigned int max_dimension = std::numeric_limits<unsigned int>::max(), bool check_euler_characteristic = true) {
 		compute_zeroth_persistence(min_dimension, max_dimension);
 		compute_higher_persistence(min_dimension, max_dimension);
 		complex.finished();
 		output->finished(check_euler_characteristic);
 
 		// Sanity check whether there were any problems computing the homology
-		bool computed_full_homology = min_dimension == 0 && max_dimension == std::numeric_limits<unsigned short>::max();
+		bool computed_full_homology = min_dimension == 0 && max_dimension == std::numeric_limits<unsigned int>::max();
 		if (check_euler_characteristic && computed_full_homology && max_entries == std::numeric_limits<size_t>::max()) {
 			index_t cell_euler_characteristic = 0;
 			for (size_t i = 0; i <= complex.top_dimension(); i++) {
@@ -527,7 +527,7 @@ public:
 #endif
 
 protected:
-	void compute_zeroth_persistence(unsigned short min_dimension, unsigned short) {
+	void compute_zeroth_persistence(unsigned int min_dimension, unsigned int) {
 		complex.prepare_next_dimension(0);
 
 		// Only compute this if we actually need it
@@ -560,7 +560,7 @@ protected:
 
 		for (auto e : edges) {
 			const auto vertices = complex.vertices_of_edge(get_index(e));
-			index_t u = dset.find(vertices.first), v = dset.find(vertices.second);	
+			index_t u = dset.find(vertices.first), v = dset.find(vertices.second);
 
 			if (u != v) {
 				// Only output bars if we are interested in zeroth homology
@@ -619,7 +619,7 @@ protected:
 		euler_characteristic += betti_number;
 	}
 
-	void compute_higher_persistence(unsigned short min_dimension, unsigned short max_dimension) {
+	void compute_higher_persistence(unsigned int min_dimension, unsigned int max_dimension) {
 		for (index_t dimension = 1; dimension <= max_dimension; ++dimension) {
 			if (dimension + 1 == min_dimension) {
 				// Here we need to reduce *all* cells because we did not compute anything of smaller dimension
